refactor(api): extract InterfaceFlags so InterfaceInfo matches the Info/Flags family
Mirror SessionInfo/SessionFlags: InterfaceInfo becomes a record exposing an
InterfaceFlags record (ForkSafe) instead of inline Flags + IsForkSafe. Adds the
CKF_INTERFACE_FORK_SAFE constant (replacing a 0x1 literal) and unit tests for
both new types (100% line/branch). Pre-ship; no compatibility concern.

+++ src/KerckhoffsLabs.Security.Cryptography.Pkcs11/Pkcs11Library.cs
@@ -244,7 +244,7 @@ public sealed class Pkcs11Library : IDisposable
             string name = raw[i].InterfaceName != IntPtr.Zero
                 ? Marshal.PtrToStringUTF8(raw[i].InterfaceName) ?? string.Empty
                 : string.Empty;
-            list.Add(new InterfaceInfo(name, (ulong)raw[i].Flags));
+            list.Add(new InterfaceInfo(name, raw[i].Flags));
         }
 
         return list;
